requirements.txt
langchain
pypdf2
loguru
streamlit
streamlit_chatbot
python-dotenv
openai
# faiss-cpu
# tiktoken
chromadb
